+++ leshan-core/src/main/java/org/eclipse/leshan/core/request/DownLinkRequestVisitorAdapter.java
@@ -16,8 +16,6 @@
  *******************************************************************************/
 package org.eclipse.leshan.core.request;
 
-import org.eclipse.leshan.core.request.argument.ExecuteRequest;
-
 /**
  * An abstract adapter class for visiting down link request. The methods in this class are empty. This class exists as
  * convenience for creating message observer objects.
